rename: reward 관련 internal dto 위치 변경
