+++ src/main/java/com/grepp/spring/app/controller/api/reward/payload/RewardItemResponse.java
@@ -1,6 +1,6 @@
 package com.grepp.spring.app.controller.api.reward.payload;
 
-import com.grepp.spring.app.model.reward.dto.RewardItemDto;
+import com.grepp.spring.app.model.reward.dto.internal.RewardItemDto;
 import java.util.List;
 
 public record RewardItemResponse(
