+++ src/main/java/com/grepp/spring/app/model/reward/dto/internal/OwnItemDto.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.model.reward.dto;
+package com.grepp.spring.app.model.reward.dto.internal;
 
 import com.grepp.spring.app.model.reward.code.ItemType;
 import com.querydsl.core.annotations.QueryProjection;
