+++ src/main/java/com/grepp/spring/app/model/reward/service/OwnItemService.java
@@ -4,8 +4,8 @@ package com.grepp.spring.app.model.reward.service;
 import com.grepp.spring.app.model.member.entity.Member;
 import com.grepp.spring.app.model.member.repository.MemberRepository;
 import com.grepp.spring.app.model.reward.code.ItemType;
-import com.grepp.spring.app.model.reward.dto.ItemSetDto;
-import com.grepp.spring.app.model.reward.dto.OwnItemDto;
+import com.grepp.spring.app.model.reward.dto.internal.ItemSetDto;
+import com.grepp.spring.app.model.reward.dto.internal.OwnItemDto;
 import com.grepp.spring.app.model.reward.entity.OwnItem;
 import com.grepp.spring.app.model.reward.entity.RewardItem;
 import com.grepp.spring.app.model.reward.repository.OwnItemRepository;
